UIPFU-28: Can't filter requester lookup by patron group with commas
Addressed this by defining a separate displayName in the FiterGroup
values when they are mapped on component load, which displays the
unmodified user group name.  Then wrapped a javascript encodeURIComponent
function around the internal identifier group name
to encode any special characters (such as spaces and commas).
This seems to have fixed the problem, user groups with commas in the
names can now be selected normally and will pull up user names in the
modal.

Refs: https://issues.folio.org/browse/UIPFU-28